[FIX] project: format allocated time without hr_timesheet
Before this commit, opening the portal page of a task in a database
without hr_timesheet answers a 500:

    AttributeError: 'account.analytic.line' object has no attribute
    '_format_portal_hours'

This happens because project's portal_my_task_allocated_hours_template
formats the allocated time with `_format_portal_hours`, a method
hr_timesheet adds on `account.analytic.line`.

This commit moves that method to project, next to the template that
calls it.

https://runbot.odoo.com/odoo/error/946944
