+++ addons/project/models/__init__.py
@@ -2,6 +2,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import account_analytic_account
+from . import account_analytic_line
 from . import mail_activity
 from . import mail_message
 from . import project_project_stage
